fix: ajusta validação da rota de deleção de usuário

+++ app.py
@@ -92,7 +92,7 @@ def update_user(id_user):
 @login_required
 def delete_user(id_user):
     user = User.query.get(id_user)
-    if user:
+    if user and id_user != current_user.id:
         db.session.delete(user)
         db.session.commit()
         return jsonify({"message": f"Usuário {id_user} deletado com sucesso"})
